The script exits quickly in most cases
Previously, it could happen that the decoupled-connect script did not notice that the two processes it connects are done, which meant that it kept running for an unnecessarily long time. This is fixed now.

The script also provides more debug output aboug what it did, including exit codes of the terminated processes.